requirements.txt
# backend heavy----
# flask
# flask-cors
# tensorflow
# numpy
# opencv-python
# Pillow
# gunicorn


flask
flask-cors
tensorflow
numpy
opencv-python-headless
Pillow
gunicorn